requirements.txt
astroid==2.4.1
attrs==19.3.0
backcall==0.1.0
bleach==3.1.5
Brotli==1.0.7
click==7.1.2
dash==1.12.0
dash-core-components==1.10.0
dash-html-components==1.0.3
dash-renderer==1.4.1
dash-table==4.7.0
decorator==4.4.2
defusedxml==0.6.0
dominate==2.5.1
entrypoints==0.3
Flask==1.0.2
Flask-Bootstrap4==4.0.2
Flask-Compress==1.5.0
Flask-Cors==3.0.4
Flask-Login==0.4.1
Flask-Material==0.1.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
future==0.18.2
importlib-metadata==1.6.0
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.4
packaging==20.3
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
plotly==4.7.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pylint==2.5.2
PyMySQL==0.9.3
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.1
retrying==1.3.3
Send2Trash==1.5.0
six==1.14.0
SQLAlchemy==1.3.17
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.9
visitor==0.1.3
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.2.1
wrapt==1.12.1
WTForms==2.2.1
zipp==3.1.0
